requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.13
django-dump-die==0.1.7
django-vite==3.0.4
djangorestframework==3.15.2
executing==2.0.1
idna==3.7
inertia-django==0.6.0
Pygments==2.18.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
